refactor: initial structure
