+++ src/errors.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum GameError {
+    RequiresFourReadyPlayers,
+    UnbalancedTeams,
+}
