+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod user;
 mod game;
+mod errors;
